requirements.txt
# You can add your project dependencies here, for example:
# django==5.0.2
# requests==2.31.0 

# Web Framework
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
streamlit==1.29.0

# Data Processing and Analysis
numpy>=1.22.4,<2.0.0
pandas==2.1.4
geopandas==0.14.1
shapely==2.0.2

# Visualization
plotly==5.18.0
pydeck==0.8.0

# Map Services
mapboxgl==0.10.2

# GIS and Spatial Data
pyproj==3.6.1
fiona==1.9.5
rtree==1.1.0

# Database
sqlalchemy==2.0.23

# Validation and Testing
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0

# Security
python-dotenv==1.0.0
flask-login==0.6.3
flask-wtf==1.2.1

# API and Data Format Support
xmltodict==0.13.0
geojson==3.1.0

# Utils
tqdm==4.66.1
click==8.1.7
python-dateutil==2.8.2

# Optional: Install these after core dependencies if needed
# rasterio==1.3.9  # Commented out due to numpy conflicts
# cesium==0.12.3   # Commented out due to numpy conflicts